Try harder to find dash installed docsets
The paths where dash itself installs the docsets is somewhat
unpredictable (#84) and can not be reliably determined from the docset
name/id alone. For example:
* Bootstrap_2 => Bootstrap_2/Bootstrap.docset
* Bootstrap_3 => Bootstrap_3/Bootstrap 3.docset
* Emacs_Lisp => Emacs_Lisp/Emacs Lisp.docset

Try harder to find them simply by looking inside the directory and
accepting the first *.docset directory present. The list of paths
searched is now:
* {name/id}.docset             -- helm-dash installation, probably
* {name/id}/{name/id}.docset   -- dash installation, e.g. Haskell
* {name/id}/*.docset           -- dash installation, e.g. Bootstrap_2