Rename `helm-dash-sql-execute` to `helm-dash-sql-query` and update it
This method isn't executing sql commands, it's returning a sql query to
execute, so this is a proper name for it.

Also, as helm-dash only gets content, it doesn't make sense to have "query-type"
as an option, it's always "select". I update it also to don't crash if a invalid
docset type is sent.