Update install functions to work with a docset name argument
These functions were extracting the content of the tar.gz file and taking the
docset name from the extracted folder. This was limiting us because, for
example, Dash has 3 different docsets for Lua (Lua_5.1, Lua_5.2, Lua_5.3) but
the folder name of all of them would be just "Lua".  With this change, we're
replicating exactly the same directory structure of Dash app for official
docsets. Ex: `.../Lua_5.1/Lua.docset/.../