Update travis config file
Remove `sudo: false` option. Looks like with that option, travisci is sharing
IPs and our tests are exceeding Github rate limit. If we would like to fix it
properly, we would need to update our code to call github using personal tokens
and update tests to generate & use a token. ATM, it doesn't worth implementing
that.