Do not use the docset tarball top directory as docset name.
In some newer docsets (Ruby_2 for example), the top directory of the docset
archive (Ruby.docset) does not match the docset name. It breaks the docset
detection has helm-dash is looking for Ruby_2.docset instead of Ruby.docest. It
also prevents having docsets with different names but with the same top
folders (Ruby and Ruby_2 for example) to coexists.

Install docset in their matching directory by stripping the top level directory
from the archive when extracting to the docset content directory.